+++ datumaro/components/annotation.py
@@ -249,9 +249,10 @@ class Mask(Annotation):
 
     @property
     def image(self) -> BinaryMaskImage:
-        if callable(self._image):
-            return self._image()
-        return self._image
+        image = self._image
+        if callable(image):
+            image = image()
+        return image
 
     def as_class_mask(self, label_id: Optional[int] = None) -> IndexMaskImage:
         """
@@ -304,16 +305,25 @@ class RleMask(Mask):
     An RLE-encoded instance segmentation mask.
     """
 
-    rle = field() # uses pycocotools RLE representation
+    _rle = field() # uses pycocotools RLE representation
 
-    _image = field(init=False, default=attr.Factory(
-        lambda self: self._lazy_decode(self.rle),
-        takes_self=True))
+    _image = field(init=False, default=None)
+
+    @property
+    def image(self) -> BinaryMaskImage:
+        return self._decode(self.rle)
+
+    @property
+    def rle(self):
+        rle = self._rle
+        if callable(rle):
+            rle = rle()
+        return rle
 
     @staticmethod
-    def _lazy_decode(rle):
+    def _decode(rle):
         from pycocotools import mask as mask_utils
-        return lambda: mask_utils.decode(rle)
+        return mask_utils.decode(rle)
 
     def get_area(self) -> int:
         from pycocotools import mask as mask_utils